JC-165: Assign errors to a field on periods pages
Play's form library makes cross-field validation difficult to do well.
Out of the box, you have to apply a global constraint - but this means
the resulting error is at the global scope.  We need to be able to assign
errors to a field, so we can help the user fix the issue.

This solution isn't ideal, but uses a specific Formatter because that's
one of the few places in code that we can access the values of other
fields in the form.  Most of the cross-field errors could be because
of a mistake in either the start or end date, but the wording of the
errors means that assigning them to the end date makes most sense.